refactor(orchestration): move assignCourtToGroup to facade layer
WP5.5 orchestration extraction - Move assignCourtToGroup (~368 lines)
to dedicated orchestrator file. This is the LARGEST orchestrator in the
codebase. No logic changes - verbatim move with deps injection.

DEPENDENCY CHECKLIST (call-site parity):

Read values:
- isAssigning, mobileFlow, preselectedCourt, operatingHours
- currentGroup, courts (data.courts), currentWaitlistEntryId, CONSTANTS

Setters (16 total):
- setIsAssigning, setCurrentWaitlistEntryId, setHasWaitlistPriority
- setCurrentGroup, setJustAssignedCourt, setAssignedSessionId
- setReplacedGroup, setDisplacement, setOriginalCourtData
- setIsChangingCourt, setWasOvertimeCourt, setHasAssignedCourt
- setCanChangeCourt, setChangeTimeRemaining, setIsTimeLimited
- setTimeLimitReason, setShowSuccess, setGpsFailedPrompt

Services:
- backend.commands.assignFromWaitlist
- backend.commands.assignCourtWithPlayers
- backend.queries.refresh

Helpers:
- getCourtBlockStatus, getMobileGeolocation, showAlertMessage
- validateGroupCompat, clearSuccessResetTimer, resetForm
- successResetTimerRef, dbg, API_CONFIG

App.jsx retains thin wrapper passing deps to orchestrated function.
243 unit tests pass, 15 e2e tests pass.
